Rename model.py to core.py
